Reorganized date extension
Moved date extension to a new folder called extensions
Removed the date formatting extension form the string class
Added date formatting extension to the Date class
